Updated
Added gym page terms that are created upon registration of the page. 
Pulled gym page terms from the wordpress database to be echoed into the page with php. 
Made the direction change of making the page editable on the same page, which is displayed for users who are the author of that page.